Move TimeSeriesNode to graph components module.
