+++ cai_causal_graph/graph_components.py
@@ -20,7 +20,8 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 
 from cai_causal_graph.exceptions import CausalGraphErrors
 from cai_causal_graph.interfaces import CanDictSerialize, HasIdentifier, HasMetadata
-from cai_causal_graph.type_definitions import EDGE_T, NodeLike, NodeVariableType
+from cai_causal_graph.type_definitions import EDGE_T, TIME_LAG, VARIABLE_NAME, NodeLike, NodeVariableType
+from cai_causal_graph.utils import get_name_with_lag, get_variable_name_and_lag
 
 
 class Node(HasIdentifier, HasMetadata, CanDictSerialize):
@@ -34,7 +35,7 @@ class Node(HasIdentifier, HasMetadata, CanDictSerialize):
     ):
         """
         :param identifier: String that uniquely identifies the node within the causal graph.
-        :param meta: The meta values for the node.
+        :param meta: The metadata of the node. Default is None.
         :param variable_type: The variable type that the node represents. The choices are available through the
             `cai_causal_graph.type_definitions.NodeVariableType` enum.
         """
@@ -189,6 +190,79 @@ class Node(HasIdentifier, HasMetadata, CanDictSerialize):
             node_dict['meta'] = deepcopy(self.meta)   # type: ignore
 
         return node_dict
+
+
+class TimeSeriesNode(Node):
+    """
+    Time series node.
+
+    A node in a time series causal graph will have additional metadata and attributes that gives the time information
+    of the node together with the variable name.
+
+    The two additional metadata are:
+    - `cai_causal_graph.type_definitions.TIME_LAG`: the time difference with respect to the reference time 0
+    - `cai_causal_graph.type_definitions.VARIABLE_NAME`: the name of the variable (without the lag information)
+    """
+
+    def __init__(
+        self,
+        identifier: Optional[NodeLike] = None,
+        time_lag: Optional[int] = None,
+        variable_name: Optional[str] = None,
+        meta: Optional[Dict[str, Any]] = None,
+        variable_type: NodeVariableType = NodeVariableType.UNSPECIFIED,
+    ):
+        """
+        Initialize the time series node.
+
+        :param identifier: String that uniquely identifies the node within the causal graph. If the `identifier` is
+            provided, the `time_lag` and `variable_name` will be extracted from the `identifier`. Default is `None`.
+        :param time_lag: The time lag of the node. If `time_lag` is provided, then `variable_name` must be provided
+            to set the identifier. If both `time_lag` and `variable_name` are provided, the `identifier` must be `None`.
+            Default is `None`.
+        :param variable_name: The variable name of the node. If `variable_name` is provided, then `time_lag` must be
+            provided to set the identifier. If both `time_lag` and `variable_name` are provided, the `identifier` must
+            be `None`. Default is `None`.
+        :param meta: The metadata of the node. Default is None.
+        :param variable_type: The variable type that the node represents. The choices are available through the
+            `cai_causal_graph.type_definitions.NodeVariableType` enum.
+        """
+        if time_lag is not None and variable_name is not None:
+            assert identifier is None, 'If `time_lag` and `variable_name` are provided, `identifier` must be `None`.'
+            identifier = get_name_with_lag(variable_name, time_lag)
+        elif identifier is not None:
+            assert (
+                time_lag is None and variable_name is None
+            ), 'If `identifier` is provided, `time_lag` and `variable_name` must be `None`.'
+            identifier = Node.identifier_from(identifier)
+            variable_name, time_lag = get_variable_name_and_lag(identifier)
+        else:
+            raise ValueError(
+                'Either `identifier` or both `time_lag` and `variable_name` must be provided to initialize a time '
+                'series node.'
+            )
+
+        # populate the metadata for each node
+        if meta is not None:
+            meta.update({TIME_LAG: time_lag, VARIABLE_NAME: variable_name})
+        else:
+            meta = {TIME_LAG: time_lag, VARIABLE_NAME: variable_name}
+
+        # populate the metadata for the node
+        super().__init__(identifier, meta, variable_type)
+
+    @property
+    def time_lag(self) -> int:
+        """Return the time lag of the node from the metadata."""
+        return self.meta.get(TIME_LAG, 0)
+
+    @property
+    def variable_name(self) -> Optional[str]:
+        """Return the variable name of the node from the metadata."""
+        return self.meta.get(VARIABLE_NAME, None)
+
+    def __repr__(self) -> str:
+        return super().__repr__().replace('Node', 'TimeSeriesNode')
 
 
 class Edge(HasIdentifier, HasMetadata, CanDictSerialize):
